clean up options
