+++ crates/oxc_formatter/examples/formatter.rs
@@ -5,7 +5,7 @@ use std::{fs, path::Path};
 use pico_args::Arguments;
 
 use oxc_allocator::Allocator;
-use oxc_formatter::{Formatter, FormatterOptions};
+use oxc_formatter::{FormatOptions, Formatter};
 use oxc_parser::{ParseOptions, Parser};
 use oxc_span::SourceType;
 
@@ -27,7 +27,7 @@ fn main() -> Result<(), String> {
         println!("Parsed with Errors.");
     }
 
-    let options = FormatterOptions::default();
+    let options = FormatOptions::default();
     let code = Formatter::new(&allocator, options).build(&ret.program);
 
     println!("{code}");
