+++ crates/oxc_formatter/src/formatter/printer/printer_options/mod.rs
@@ -1,4 +1,5 @@
-use super::super::{FormatOptions, IndentStyle, IndentWidth, LineEnding, LineWidth};
+use crate::FormatOptions;
+use crate::options::{IndentStyle, IndentWidth, LineEnding, LineWidth};
 
 /// Options that affect how the [crate::Printer] prints the format tokens
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -44,16 +45,13 @@ impl From<PrintWidth> for usize {
     }
 }
 
-impl<'a, O> From<&'a O> for PrinterOptions
-where
-    O: FormatOptions,
-{
-    fn from(options: &'a O) -> Self {
+impl<'a> From<&'a FormatOptions> for PrinterOptions {
+    fn from(options: &'a FormatOptions) -> Self {
         PrinterOptions::default()
-            .with_indent_style(options.indent_style())
-            .with_indent_width(options.indent_width())
-            .with_print_width(options.line_width().into())
-            .with_line_ending(options.line_ending())
+            .with_indent_style(options.indent_style)
+            .with_indent_width(options.indent_width)
+            .with_print_width(options.line_width.into())
+            .with_line_ending(options.line_ending)
     }
 }
