+++ crates/oxc_formatter/src/lib.rs
@@ -15,7 +15,6 @@
 mod generated {
     pub mod format;
 }
-mod context;
 mod formatter;
 mod options;
 mod write;
@@ -23,25 +22,25 @@ mod write;
 use oxc_allocator::Allocator;
 use oxc_ast::ast::Program;
 
-use crate::{context::JsFormatOptions, formatter::FormatContext};
+use crate::formatter::FormatContext;
 
 pub use crate::options::*;
 
 pub struct Formatter<'a> {
     allocator: &'a Allocator,
     source_text: &'a str,
-    options: FormatterOptions,
+    options: FormatOptions,
 }
 
 impl<'a> Formatter<'a> {
-    pub fn new(allocator: &'a Allocator, options: FormatterOptions) -> Self {
+    pub fn new(allocator: &'a Allocator, options: FormatOptions) -> Self {
         Self { allocator, source_text: "", options }
     }
 
     pub fn build(&mut self, program: &Program<'a>) -> String {
         let source_text = program.source_text;
         self.source_text = source_text;
-        let options = JsFormatOptions::default();
+        let options = FormatOptions::default();
         let context = FormatContext::new(program, options);
         let formatted = formatter::format(
             program,
