+++ crates/oxc_formatter/src/options.rs
@@ -1,252 +1,848 @@
-use std::str::FromStr;
+use std::marker::PhantomData;
+use std::{fmt, num::ParseIntError, rc::Rc, str::FromStr};
 
-/// Formatter Options
-///
-/// References
-/// * <https://prettier.io/docs/en/options>
-/// * <https://github.com/prettier/prettier/blob/3.3.3/src/main/core-options.evaluate.js>
-/// * <https://github.com/prettier/prettier/blob/3.3.3/src/language-js/options.js>
-#[derive(Debug, Clone, Copy)]
-pub struct FormatterOptions {
-    /* Global Options */
-    /// Specify the line length that the printer will wrap on.
-    /// Default: 80
-    pub print_width: usize,
+use oxc_allocator::Vec;
+use oxc_ast::{Comment, ast::Program};
 
-    /// Specify the number of spaces per indentation-level.
-    /// Default: 2
-    pub tab_width: usize,
+use crate::formatter::formatter::Formatter;
+use crate::formatter::prelude::{dynamic_text, if_group_breaks, text};
+use crate::formatter::printer::PrinterOptions;
+pub use crate::formatter::{
+    Buffer, Format, FormatResult, SourceComment, SyntaxTriviaPieceComments, token::string::Quote,
+};
+use crate::write;
 
-    /// Indent lines with tabs instead of spaces.
-    pub use_tabs: bool,
+// TODO: rename these to align with prettier
+#[derive(Debug, Default, Clone)]
+pub struct FormatOptions {
+    /// The indent style.
+    pub indent_style: IndentStyle,
 
-    /// Which end of line characters to apply.
-    /// Default: "lf"
-    pub end_of_line: EndOfLine,
+    /// The indent width.
+    pub indent_width: IndentWidth,
 
-    /* JavaScript Options */
-    /// Print semicolons at the ends of statements.
-    /// Default: true
-    pub semi: bool,
+    /// The type of line ending.
+    pub line_ending: LineEnding,
 
-    /// Use single quotes instead of double quotes.
-    /// Default: false
-    pub single_quote: bool,
+    /// What's the max width of a line. Defaults to 80.
+    pub line_width: LineWidth,
 
-    /// Change when properties in objects are quoted.
-    /// Default: [QuoteProps::AsNeeded]
-    pub quote_props: QuoteProps,
+    /// The style for quotes. Defaults to double.
+    pub quote_style: QuoteStyle,
 
-    /// Use single quotes instead of double quotes in JSX.
-    /// Default: false
-    pub jsx_single_quote: bool,
+    /// The style for JSX quotes. Defaults to double.
+    pub jsx_quote_style: QuoteStyle,
 
-    /// Print trailing commas wherever possible in multi-line comma-separated syntactic
-    /// structures. (A single-line array, for example, never gets trailing commas.)
-    /// Default: [TrailingComma::All]
-    pub trailing_comma: TrailingComma,
+    /// When properties in objects are quoted. Defaults to as-needed.
+    pub quote_properties: QuoteProperties,
 
-    /// Print spaces between brackets in object literals.
-    ///
-    /// * true - Example: `{ foo: bar }`.
-    /// * false - Example: `{foo: bar}`.
-    ///
-    /// Default: true
-    pub bracket_spacing: bool,
+    /// Print trailing commas wherever possible in multi-line comma-separated syntactic structures. Defaults to "all".
+    pub trailing_commas: TrailingCommas,
 
-    /// Configure how wraps object literals when they could fit on one line or span multiple lines.
-    /// Default: [ObjectWrap::Preserve]
-    pub object_wrap: ObjectWrap,
+    /// Whether the formatter prints semicolons for all statements, class members, and type members or only when necessary because of [ASI](https://tc39.es/ecma262/multipage/ecmascript-language-lexical-grammar.html#sec-automatic-semicolon-insertion).
+    pub semicolons: Semicolons,
 
-    /// Put the `>` of a multi-line HTML (HTML, JSX) element at the end of the last line
-    /// instead of being alone on the next line (does not apply to self closing elements).
-    /// Default: false
-    pub bracket_same_line: bool,
+    /// Whether to add non-necessary parentheses to arrow functions. Defaults to "always".
+    pub arrow_parentheses: ArrowParentheses,
 
-    /// Include parentheses around a sole arrow function parameter.
-    /// Default: [ArrowParens::Always]
-    pub arrow_parens: ArrowParens,
+    /// Whether to insert spaces around brackets in object literals. Defaults to true.
+    pub bracket_spacing: BracketSpacing,
+
+    /// Whether to hug the closing bracket of multiline HTML/JSX tags to the end of the last line, rather than being alone on the following line. Defaults to false.
+    pub bracket_same_line: BracketSameLine,
+
+    /// Attribute position style. By default auto.
+    pub attribute_position: AttributePosition,
+
+    /// Whether to expand object and array literals to multiple lines. Defaults to "auto".
+    pub expand: Expand,
 }
 
-impl Default for FormatterOptions {
-    fn default() -> Self {
+impl FormatOptions {
+    pub fn new() -> Self {
         Self {
-            print_width: 80,
-            tab_width: 2,
-            use_tabs: false,
-            end_of_line: EndOfLine::default(),
-            semi: true,
-            single_quote: false,
-            quote_props: QuoteProps::default(),
-            jsx_single_quote: false,
-            trailing_comma: TrailingComma::default(),
-            bracket_spacing: true,
-            object_wrap: ObjectWrap::default(),
-            bracket_same_line: false,
-            arrow_parens: ArrowParens::default(),
+            indent_style: IndentStyle::default(),
+            indent_width: IndentWidth::default(),
+            line_ending: LineEnding::default(),
+            line_width: LineWidth::default(),
+            quote_style: QuoteStyle::default(),
+            jsx_quote_style: QuoteStyle::default(),
+            quote_properties: QuoteProperties::default(),
+            trailing_commas: TrailingCommas::default(),
+            semicolons: Semicolons::default(),
+            arrow_parentheses: ArrowParentheses::default(),
+            bracket_spacing: BracketSpacing::default(),
+            bracket_same_line: BracketSameLine::default(),
+            attribute_position: AttributePosition::default(),
+            expand: Expand::default(),
         }
+    }
+
+    pub fn as_print_options(&self) -> PrinterOptions {
+        PrinterOptions::from(self)
     }
 }
 
-#[derive(Debug, Clone, Copy, Default)]
-pub enum EndOfLine {
+impl fmt::Display for FormatOptions {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        writeln!(f, "Indent style: {}", self.indent_style)?;
+        writeln!(f, "Indent width: {}", self.indent_width.value())?;
+        writeln!(f, "Line ending: {}", self.line_ending)?;
+        writeln!(f, "Line width: {}", self.line_width.value())?;
+        writeln!(f, "Quote style: {}", self.quote_style)?;
+        writeln!(f, "JSX quote style: {}", self.jsx_quote_style)?;
+        writeln!(f, "Quote properties: {}", self.quote_properties)?;
+        writeln!(f, "Trailing commas: {}", self.trailing_commas)?;
+        writeln!(f, "Semicolons: {}", self.semicolons)?;
+        writeln!(f, "Arrow parentheses: {}", self.arrow_parentheses)?;
+        writeln!(f, "Bracket spacing: {}", self.bracket_spacing.value())?;
+        writeln!(f, "Bracket same line: {}", self.bracket_same_line.value())?;
+        writeln!(f, "Attribute Position: {}", self.attribute_position)?;
+        writeln!(f, "Expand lists: {}", self.expand)
+    }
+}
+
+#[derive(Debug, Default, Clone, Copy, Eq, Hash, PartialEq)]
+pub enum IndentStyle {
+    /// Tab
     #[default]
-    Lf,
-    Crlf,
-    Cr,
-    Auto,
+    Tab,
+    /// Space
+    Space,
 }
 
-impl EndOfLine {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Crlf => "\r\n",
-            Self::Cr => "\r",
-            Self::Lf | Self::Auto => "\n",
-        }
+impl IndentStyle {
+    pub const DEFAULT_SPACES: u8 = 2;
+
+    /// Returns `true` if this is an [IndentStyle::Tab].
+    pub const fn is_tab(&self) -> bool {
+        matches!(self, IndentStyle::Tab)
+    }
+
+    /// Returns `true` if this is an [IndentStyle::Space].
+    pub const fn is_space(&self) -> bool {
+        matches!(self, IndentStyle::Space)
     }
 }
 
-impl FromStr for EndOfLine {
-    type Err = ();
+impl FromStr for IndentStyle {
+    type Err = &'static str;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "crlf" => Self::Crlf,
-            "cr" => Self::Cr,
-            "auto" => Self::Auto,
-            "lf" => Self::Lf,
-            _ => Self::default(),
-        })
+        match s {
+            "tab" => Ok(Self::Tab),
+            "space" => Ok(Self::Space),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Unsupported value for this option"),
+        }
     }
 }
 
-#[derive(Debug, Clone, Copy, Default)]
-pub enum QuoteProps {
-    /// Only add quotes around object properties where required.
+impl fmt::Display for IndentStyle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            IndentStyle::Tab => std::write!(f, "Tab"),
+            IndentStyle::Space => std::write!(f, "Space"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Default)]
+pub enum LineEnding {
+    ///  Line Feed only (\n), common on Linux and macOS as well as inside git repos
+    #[default]
+    Lf,
+
+    /// Carriage Return + Line Feed characters (\r\n), common on Windows
+    Crlf,
+
+    /// Carriage Return character only (\r), used very rarely
+    Cr,
+}
+
+impl LineEnding {
+    #[inline]
+    pub const fn as_str(&self) -> &'static str {
+        match self {
+            LineEnding::Lf => "\n",
+            LineEnding::Crlf => "\r\n",
+            LineEnding::Cr => "\r",
+        }
+    }
+
+    /// Returns `true` if this is a [LineEnding::Lf].
+    pub const fn is_line_feed(&self) -> bool {
+        matches!(self, LineEnding::Lf)
+    }
+
+    /// Returns `true` if this is a [LineEnding::Crlf].
+    pub const fn is_carriage_return_line_feed(&self) -> bool {
+        matches!(self, LineEnding::Crlf)
+    }
+
+    /// Returns `true` if this is a [LineEnding::Cr].
+    pub const fn is_carriage_return(&self) -> bool {
+        matches!(self, LineEnding::Cr)
+    }
+}
+
+impl FromStr for LineEnding {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "lf" => Ok(Self::Lf),
+            "crlf" => Ok(Self::Crlf),
+            "cr" => Ok(Self::Cr),
+            _ => Err("Value not supported for LineEnding"),
+        }
+    }
+}
+
+impl fmt::Display for LineEnding {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            LineEnding::Lf => std::write!(f, "LF"),
+            LineEnding::Crlf => std::write!(f, "CRLF"),
+            LineEnding::Cr => std::write!(f, "CR"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Eq, Hash, PartialEq)]
+pub struct IndentWidth(u8);
+
+impl IndentWidth {
+    pub const MIN: u8 = 0;
+
+    pub const MAX: u8 = 24;
+
+    /// Return the numeric value for this [IndentWidth]
+    pub fn value(&self) -> u8 {
+        self.0
+    }
+}
+
+impl Default for IndentWidth {
+    fn default() -> Self {
+        Self(2)
+    }
+}
+
+impl FromStr for IndentWidth {
+    type Err = ParseFormatNumberError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let value = u8::from_str(s).map_err(ParseFormatNumberError::ParseError)?;
+        let value = Self::try_from(value).map_err(ParseFormatNumberError::TryFromU8Error)?;
+        Ok(value)
+    }
+}
+
+impl TryFrom<u8> for IndentWidth {
+    type Error = IndentWidthFromIntError;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        if (Self::MIN..=Self::MAX).contains(&value) {
+            Ok(Self(value))
+        } else {
+            Err(IndentWidthFromIntError(value))
+        }
+    }
+}
+
+impl fmt::Display for IndentWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let value = self.value();
+        f.write_str(&std::format!("{value}"))
+    }
+}
+
+impl fmt::Debug for IndentWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        fmt::Display::fmt(self, f)
+    }
+}
+
+/// Validated value for the `line_width` formatter options
+///
+/// The allowed range of values is 1..=320
+#[derive(Clone, Copy, Eq, PartialEq)]
+pub struct LineWidth(u16);
+
+impl LineWidth {
+    /// Minimum allowed value for a valid [LineWidth]
+    pub const MIN: u16 = 1;
+    /// Maximum allowed value for a valid [LineWidth]
+    pub const MAX: u16 = 320;
+
+    /// Return the numeric value for this [LineWidth]
+    pub fn value(&self) -> u16 {
+        self.0
+    }
+}
+
+impl Default for LineWidth {
+    fn default() -> Self {
+        Self(80)
+    }
+}
+
+impl fmt::Display for LineWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let value = self.value();
+        f.write_str(&std::format!("{value}"))
+    }
+}
+
+impl fmt::Debug for LineWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        fmt::Display::fmt(self, f)
+    }
+}
+
+/// Error type returned when parsing a [LineWidth] or [IndentWidth] from a string fails
+pub enum ParseFormatNumberError {
+    /// The string could not be parsed to a number
+    ParseError(ParseIntError),
+    /// The `u16` value of the string is not a valid [LineWidth]
+    TryFromU16Error(LineWidthFromIntError),
+    /// The `u8 value of the string is not a valid [IndentWidth]
+    TryFromU8Error(IndentWidthFromIntError),
+}
+
+impl From<IndentWidthFromIntError> for ParseFormatNumberError {
+    fn from(value: IndentWidthFromIntError) -> Self {
+        Self::TryFromU8Error(value)
+    }
+}
+
+impl From<LineWidthFromIntError> for ParseFormatNumberError {
+    fn from(value: LineWidthFromIntError) -> Self {
+        Self::TryFromU16Error(value)
+    }
+}
+
+impl From<ParseIntError> for ParseFormatNumberError {
+    fn from(value: ParseIntError) -> Self {
+        Self::ParseError(value)
+    }
+}
+
+impl fmt::Debug for ParseFormatNumberError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        fmt::Display::fmt(self, f)
+    }
+}
+
+impl fmt::Display for ParseFormatNumberError {
+    fn fmt(&self, fmt: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ParseFormatNumberError::ParseError(err) => fmt::Display::fmt(err, fmt),
+            ParseFormatNumberError::TryFromU16Error(err) => fmt::Display::fmt(err, fmt),
+            ParseFormatNumberError::TryFromU8Error(err) => fmt::Display::fmt(err, fmt),
+        }
+    }
+}
+
+impl TryFrom<u16> for LineWidth {
+    type Error = LineWidthFromIntError;
+
+    fn try_from(value: u16) -> Result<Self, Self::Error> {
+        if (Self::MIN..=Self::MAX).contains(&value) {
+            Ok(Self(value))
+        } else {
+            Err(LineWidthFromIntError(value))
+        }
+    }
+}
+
+impl FromStr for LineWidth {
+    type Err = ParseFormatNumberError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let value = u16::from_str(s).map_err(ParseFormatNumberError::ParseError)?;
+        let value = Self::try_from(value).map_err(ParseFormatNumberError::TryFromU16Error)?;
+        Ok(value)
+    }
+}
+
+/// Error type returned when converting a u16 to a [LineWidth] fails
+#[derive(Clone, Copy, Debug)]
+pub struct IndentWidthFromIntError(pub u8);
+
+impl fmt::Display for IndentWidthFromIntError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(f, "The indent width should be between {} and {}", LineWidth::MIN, LineWidth::MAX,)
+    }
+}
+
+/// Error type returned when converting a u16 to a [LineWidth] fails
+#[derive(Clone, Copy, Debug)]
+pub struct LineWidthFromIntError(pub u16);
+
+impl fmt::Display for LineWidthFromIntError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(f, "The line width should be between {} and {}", LineWidth::MIN, LineWidth::MAX,)
+    }
+}
+
+impl From<LineWidth> for u16 {
+    fn from(value: LineWidth) -> Self {
+        value.0
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum QuoteStyle {
+    #[default]
+    Double,
+    Single,
+}
+
+impl QuoteStyle {
+    pub fn from_byte(byte: u8) -> Option<QuoteStyle> {
+        match byte {
+            b'"' => Some(QuoteStyle::Double),
+            b'\'' => Some(QuoteStyle::Single),
+            _ => None,
+        }
+    }
+
+    pub fn as_char(&self) -> char {
+        match self {
+            QuoteStyle::Double => '"',
+            QuoteStyle::Single => '\'',
+        }
+    }
+
+    pub fn as_byte(&self) -> u8 {
+        self.as_char() as u8
+    }
+
+    /// Returns the quote in HTML entity
+    pub fn as_html_entity(&self) -> &str {
+        match self {
+            QuoteStyle::Double => "&quot;",
+            QuoteStyle::Single => "&apos;",
+        }
+    }
+
+    /// Given the current quote, it returns the other one
+    #[must_use]
+    pub fn other(&self) -> Self {
+        match self {
+            QuoteStyle::Double => QuoteStyle::Single,
+            QuoteStyle::Single => QuoteStyle::Double,
+        }
+    }
+
+    pub const fn is_double(&self) -> bool {
+        matches!(self, Self::Double)
+    }
+}
+
+impl FromStr for QuoteStyle {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "double" => Ok(Self::Double),
+            "single" => Ok(Self::Single),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Value not supported for QuoteStyle"),
+        }
+    }
+}
+
+impl fmt::Display for QuoteStyle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            QuoteStyle::Double => std::write!(f, "Double Quotes"),
+            QuoteStyle::Single => std::write!(f, "Single Quotes"),
+        }
+    }
+}
+
+impl From<QuoteStyle> for Quote {
+    fn from(quote: QuoteStyle) -> Self {
+        match quote {
+            QuoteStyle::Double => Self::Double,
+            QuoteStyle::Single => Self::Single,
+        }
+    }
+}
+
+#[derive(Eq, PartialEq, Debug, Copy, Clone, Hash)]
+pub struct TabWidth(u8);
+
+impl From<u8> for TabWidth {
+    fn from(value: u8) -> Self {
+        TabWidth(value)
+    }
+}
+
+impl From<TabWidth> for u8 {
+    fn from(width: TabWidth) -> Self {
+        width.0
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum QuoteProperties {
     #[default]
     AsNeeded,
-    /// If at least one property in an object requires quotes, quote all properties.
-    Consistent,
-    /// Respect the input use of quotes in object properties.
     Preserve,
 }
 
-impl QuoteProps {
-    pub fn as_needed(self) -> bool {
+impl FromStr for QuoteProperties {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "as-needed" => Ok(Self::AsNeeded),
+            "preserve" => Ok(Self::Preserve),
+            _ => Err("Value not supported for QuoteProperties"),
+        }
+    }
+}
+
+impl fmt::Display for QuoteProperties {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            QuoteProperties::AsNeeded => std::write!(f, "As needed"),
+            QuoteProperties::Preserve => std::write!(f, "Preserve"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum Semicolons {
+    #[default]
+    Always,
+    AsNeeded,
+}
+
+impl Semicolons {
+    pub const fn is_as_needed(&self) -> bool {
         matches!(self, Self::AsNeeded)
     }
 
-    pub fn preserve(self) -> bool {
-        matches!(self, Self::Preserve)
-    }
-
-    pub fn consistent(self) -> bool {
-        matches!(self, Self::Consistent)
-    }
-}
-
-impl FromStr for QuoteProps {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "as_needed" => Self::AsNeeded,
-            "consistent" => Self::Consistent,
-            "preserve" => Self::Preserve,
-            _ => Self::default(),
-        })
-    }
-}
-
-#[derive(Debug, Default, Clone, Copy, Eq, PartialEq)]
-pub enum TrailingComma {
-    /// Trailing commas wherever possible (including function parameters and calls).
-    #[default]
-    All,
-    /// Trailing commas where valid in ES5 (objects, arrays, etc.). Trailing commas in type parameters in TypeScript.
-    ES5,
-    /// No trailing commas.
-    None,
-}
-
-impl TrailingComma {
-    pub fn should_print_es5(self) -> bool {
-        self.should_print_comma_impl(false)
-    }
-
-    pub fn should_print_all(self) -> bool {
-        self.should_print_comma_impl(true)
-    }
-
-    fn should_print_comma_impl(self, level_all: bool) -> bool {
-        matches!(self, Self::All) || (matches!(self, Self::ES5) && !level_all)
-    }
-}
-
-impl FromStr for TrailingComma {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "all" => Self::All,
-            "es5" => Self::ES5,
-            "none" => Self::None,
-            _ => Self::default(),
-        })
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default)]
-pub enum ObjectWrap {
-    /// Keep as multi-line, if there is a newline between the opening brace and first property.
-    #[default]
-    Preserve,
-    /// Fit to a single line when possible.
-    Collapse,
-}
-
-impl ObjectWrap {
-    pub fn is_preserve(self) -> bool {
-        matches!(self, Self::Preserve)
-    }
-}
-
-impl FromStr for ObjectWrap {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "preserve" => Self::Preserve,
-            "collapse" => Self::Collapse,
-            _ => Self::default(),
-        })
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default)]
-pub enum ArrowParens {
-    /// Always include parens. `Example: (x) => x`
-    #[default]
-    Always,
-    /// Omit parens when possible. `Example: x => x`
-    Avoid,
-}
-
-impl ArrowParens {
-    pub fn is_always(self) -> bool {
+    pub const fn is_always(&self) -> bool {
         matches!(self, Self::Always)
     }
 }
 
-impl FromStr for ArrowParens {
-    type Err = ();
+impl FromStr for Semicolons {
+    type Err = &'static str;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "always" => Self::Always,
-            "avoid" => Self::Avoid,
-            _ => Self::default(),
-        })
+        match s {
+            "as-needed" => Ok(Self::AsNeeded),
+            "always" => Ok(Self::Always),
+            _ => Err(
+                "Value not supported for Semicolons. Supported values are 'as-needed' and 'always'.",
+            ),
+        }
+    }
+}
+
+impl fmt::Display for Semicolons {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Semicolons::AsNeeded => std::write!(f, "As needed"),
+            Semicolons::Always => std::write!(f, "Always"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum ArrowParentheses {
+    #[default]
+    Always,
+    AsNeeded,
+}
+
+impl ArrowParentheses {
+    pub const fn is_as_needed(&self) -> bool {
+        matches!(self, Self::AsNeeded)
+    }
+
+    pub const fn is_always(&self) -> bool {
+        matches!(self, Self::Always)
+    }
+}
+
+impl FromStr for ArrowParentheses {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "as-needed" => Ok(Self::AsNeeded),
+            "always" => Ok(Self::Always),
+            _ => Err(
+                "Value not supported for Arrow parentheses. Supported values are 'as-needed' and 'always'.",
+            ),
+        }
+    }
+}
+
+impl fmt::Display for ArrowParentheses {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ArrowParentheses::AsNeeded => std::write!(f, "As needed"),
+            ArrowParentheses::Always => std::write!(f, "Always"),
+        }
+    }
+}
+
+/// This enum is used within formatting functions to print or omit trailing commas.
+#[derive(Debug, Copy, Clone)]
+pub enum FormatTrailingCommas {
+    /// Print trailing commas if the option is [TrailingCommas::All].
+    All,
+    /// Print trailing commas if the option is [TrailingCommas::All] or [TrailingCommas::Es5].
+    ES5,
+}
+
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Default)]
+pub enum TrailingSeparator {
+    /// A trailing separator is allowed and preferred
+    #[default]
+    Allowed,
+
+    /// A trailing separator is not allowed
+    Disallowed,
+
+    /// A trailing separator is mandatory for the syntax to be correct
+    Mandatory,
+
+    /// A trailing separator might be present, but the consumer
+    /// decides to remove it
+    Omit,
+}
+
+impl FormatTrailingCommas {
+    /// This function returns corresponding [TrailingSeparator] for `format_separated` function.
+    pub fn trailing_separator(&self, options: &FormatOptions) -> TrailingSeparator {
+        if options.trailing_commas.is_none() {
+            return TrailingSeparator::Omit;
+        }
+
+        match self {
+            FormatTrailingCommas::All => {
+                if options.trailing_commas.is_all() {
+                    TrailingSeparator::Allowed
+                } else {
+                    TrailingSeparator::Omit
+                }
+            }
+            FormatTrailingCommas::ES5 => TrailingSeparator::Allowed,
+        }
+    }
+}
+
+impl Format<'_> for FormatTrailingCommas {
+    fn fmt(&self, f: &mut Formatter) -> FormatResult<()> {
+        if f.options().trailing_commas.is_none() {
+            return Ok(());
+        }
+
+        if matches!(self, FormatTrailingCommas::ES5) || f.options().trailing_commas.is_all() {
+            write!(f, [if_group_breaks(&text(","))])?;
+        }
+
+        Ok(())
+    }
+}
+
+/// Print trailing commas wherever possible in multi-line comma-separated syntactic structures.
+#[derive(Clone, Copy, Default, Debug, Eq, Hash, PartialEq)]
+pub enum TrailingCommas {
+    /// Trailing commas wherever possible (including function parameters and calls).
+    #[default]
+    All,
+    /// Trailing commas where valid in ES5 (objects, arrays, etc.). No trailing commas in type parameters in TypeScript.
+    Es5,
+    /// No trailing commas.
+    None,
+}
+
+impl TrailingCommas {
+    pub const fn is_es5(&self) -> bool {
+        matches!(self, TrailingCommas::Es5)
+    }
+    pub const fn is_all(&self) -> bool {
+        matches!(self, TrailingCommas::All)
+    }
+    pub const fn is_none(&self) -> bool {
+        matches!(self, TrailingCommas::None)
+    }
+}
+
+impl FromStr for TrailingCommas {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "es5" => Ok(Self::Es5),
+            "all" => Ok(Self::All),
+            "none" => Ok(Self::None),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Value not supported for TrailingCommas"),
+        }
+    }
+}
+
+impl fmt::Display for TrailingCommas {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            TrailingCommas::Es5 => std::write!(f, "ES5"),
+            TrailingCommas::All => std::write!(f, "All"),
+            TrailingCommas::None => std::write!(f, "None"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum AttributePosition {
+    #[default]
+    Auto,
+    Multiline,
+}
+
+impl fmt::Display for AttributePosition {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            AttributePosition::Auto => std::write!(f, "Auto"),
+            AttributePosition::Multiline => std::write!(f, "Multiline"),
+        }
+    }
+}
+
+impl FromStr for AttributePosition {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "multiline" => Ok(Self::Multiline),
+            "auto" => Ok(Self::Auto),
+            _ => Err(
+                "Value not supported for attribute_position. Supported values are 'auto' and 'multiline'.",
+            ),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub struct BracketSpacing(bool);
+
+impl BracketSpacing {
+    /// Return the boolean value for this [BracketSpacing]
+    pub fn value(&self) -> bool {
+        self.0
+    }
+}
+
+impl Default for BracketSpacing {
+    fn default() -> Self {
+        Self(true)
+    }
+}
+
+impl From<bool> for BracketSpacing {
+    fn from(value: bool) -> Self {
+        Self(value)
+    }
+}
+
+impl fmt::Display for BracketSpacing {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::write!(f, "{:?}", self.value())
+    }
+}
+
+impl FromStr for BracketSpacing {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let value = bool::from_str(s);
+
+        match value {
+            Ok(value) => Ok(Self(value)),
+            Err(_) => Err(
+                "Value not supported for BracketSpacing. Supported values are 'true' and 'false'.",
+            ),
+        }
+    }
+}
+
+/// Put the `>` of a multi-line HTML or JSX element at the end of the last line instead of being alone on the next line (does not apply to self closing elements).
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub struct BracketSameLine(bool);
+
+impl BracketSameLine {
+    /// Return the boolean value for this [BracketSameLine]
+    pub fn value(&self) -> bool {
+        self.0
+    }
+}
+
+impl From<bool> for BracketSameLine {
+    fn from(value: bool) -> Self {
+        Self(value)
+    }
+}
+
+impl fmt::Display for BracketSameLine {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::write!(f, "{}", self.value())
+    }
+}
+
+impl FromStr for BracketSameLine {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match bool::from_str(s) {
+            Ok(value) => Ok(Self(value)),
+            Err(_) => Err(
+                "Value not supported for BracketSameLine. Supported values are 'true' and 'false'.",
+            ),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum Expand {
+    /// Objects are expanded when the first property has a leading newline. Arrays are always
+    /// expanded if they are shorter than the line width.
+    #[default]
+    Auto,
+    /// Objects and arrays are always expanded.
+    Always,
+    /// Objects and arrays are never expanded, if they are shorter than the line width.
+    Never,
+}
+
+impl FromStr for Expand {
+    type Err = String;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "auto" => Ok(Self::Auto),
+            "always" => Ok(Self::Always),
+            "never" => Ok(Self::Never),
+            _ => Err(std::format!("unknown expand literal: {s}")),
+        }
+    }
+}
+
+impl fmt::Display for Expand {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Expand::Auto => std::write!(f, "Auto"),
+            Expand::Always => std::write!(f, "Always"),
+            Expand::Never => std::write!(f, "Never"),
+        }
     }
 }
